Create README.md
With over 20 years of experience in operations, training, transitions and having used data and worked with data at every level, I am able to ideas, problems, challenges and first help visualize and provide a perspective. Then slice & dice, break down, and sequence it. Finally transform, defend, preserve and Bring it to life or Lay it to rest.